requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
requests
python-dotenv
cryptography
python-jose[cryptography]
passlib
argon2-cffi
python-multipart
pydantic[email]
pydantic-settings
google-generativeai